Implement A-Packages: Divide Classes into packages
