+++ src/main/java/chatkaki/commands/Command.java
@@ -1,3 +1,5 @@
+package chatkaki.commands;
+
 import java.io.IOException;
 
 public abstract class Command {
