+++ src/main/java/chatkaki/commands/CommandBye.java
@@ -1,4 +1,7 @@
-import java.time.LocalDateTime;
+package chatkaki.commands;
+
+import chatkaki.Ui;
+import chatkaki.Storage;
 
 public class CommandBye extends Command {
