+++ src/main/java/chatkaki/commands/CommandEvent.java
@@ -1,3 +1,10 @@
+package chatkaki.commands;
+
+import chatkaki.tasks.TaskList;
+import chatkaki.tasks.Event;
+import chatkaki.Ui;
+import chatkaki.DateTimeHelper;
+
 import java.time.LocalDateTime;
 
 public class CommandEvent extends Command {
