+++ src/main/java/chatkaki/commands/CommandTodo.java
@@ -1,3 +1,9 @@
+package chatkaki.commands;
+
+import chatkaki.Ui;
+import chatkaki.tasks.TaskList;
+import chatkaki.tasks.Todo;
+
 public class CommandTodo extends Command {
     private String[] inputs;
