+++ src/main/java/chatkaki/commands/CommandTypes.java
@@ -1,3 +1,5 @@
+package chatkaki.commands;
+
 public enum CommandTypes {
     DELETE, TODO, DEADLINE, EVENT, BYE, MARK, UNMARK, LIST, UNKNOWN;
