+++ src/main/java/chatkaki/commands/CommandUnmark.java
@@ -1,7 +1,13 @@
-public class CommandMark extends Command {
+package chatkaki.commands;
+
+import chatkaki.Ui;
+import chatkaki.tasks.Task;
+import chatkaki.tasks.TaskList;
+
+public class CommandUnmark extends Command {
     private String[] inputs;
 
-    public CommandMark(String[] inputs) {
+    public CommandUnmark(String[] inputs) {
         this.inputs = inputs;
     }
 
@@ -13,6 +19,6 @@ public class CommandMark extends Command {
             return;
         }
         Task task = TaskList.getTask(index - 1);
-        task.markAsDone();
+        task.markAsUndone();
     }
 }
