+++ src/main/java/chatkaki/tasks/Task.java
@@ -1,3 +1,5 @@
+package chatkaki.tasks;
+
 public class Task {
     private String description;
     private boolean isDone = false;
