+++ src/main/java/chatkaki/tasks/TaskList.java
@@ -1,4 +1,7 @@
+package chatkaki.tasks;
+
 import java.util.ArrayList;
+import chatkaki.Ui;
 
 public class TaskList {
