+++ src/main/java/chatkaki/tasks/TaskType.java
@@ -1,3 +1,5 @@
+package chatkaki.tasks;
+
 public enum TaskType {
     TODO, DEADLINE, EVENT
 }
