+++ text-ui-test/runtest.sh
@@ -13,7 +13,7 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/*.java
+if ! javac -cp ../src/main/java -Xlint:none -d ../bin ../src/main/java/chatkaki/*.java ../src/main/java/chatkaki/tasks/*.java ../src/main/java/chatkaki/commands/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
